Sign RPM metadata
This mitigates vulnerabilities in librepo, and prevents an attacker from
providing a set of packages that are inconsistent with each other.  It
can also protect against downgrade attacks, although limitations in DNF
prevent it from doing so at this time.

I also massively refactored various shell scripts; this accounts for the
bulk of the changes.  Inputs are now validated, and better errors are
displayed if a script is misused.